add: base speaker controller
